+++ app/src/main/java/yokai/core/di/DomainModule.kt
@@ -9,6 +9,7 @@ import yokai.data.library.custom.CustomMangaRepositoryImpl
 import yokai.data.libraryUpdateError.LibraryUpdateErrorRepositoryImpl
 import yokai.data.manga.ExcludedScanlatorsRepositoryImpl
 import yokai.data.manga.MangaRepositoryImpl
+import yokai.data.manga.RelatedMangaCacheRepositoryImpl
 import yokai.data.source.browse.filter.SavedSearchRepositoryImpl
 import yokai.data.track.TrackRepositoryImpl
 import yokai.domain.category.CategoryRepository
@@ -42,11 +43,15 @@ import yokai.domain.library.custom.interactor.RelinkCustomManga
 import yokai.domain.libraryUpdateError.LibraryUpdateErrorRepository
 import yokai.domain.manga.ExcludedScanlatorsRepository
 import yokai.domain.manga.MangaRepository
+import yokai.domain.manga.RelatedMangaCacheRepository
 import yokai.domain.manga.interactor.GetExcludedScanlators
 import yokai.domain.manga.interactor.GetLibraryManga
 import yokai.domain.manga.interactor.GetManga
+import yokai.domain.manga.interactor.GetRelatedMangaCache
 import yokai.domain.manga.interactor.InsertManga
+import yokai.domain.manga.interactor.InvalidateRelatedMangaCache
 import yokai.domain.manga.interactor.SetExcludedScanlators
+import yokai.domain.manga.interactor.SetRelatedMangaCache
 import yokai.domain.manga.interactor.UpdateManga
 import yokai.domain.recents.interactor.GetRecents
 import yokai.domain.source.browse.filter.FilterSerializer
@@ -91,6 +96,11 @@ fun domainModule() = module {
     single<ExcludedScanlatorsRepository> { ExcludedScanlatorsRepositoryImpl(get()) }
     factory { GetExcludedScanlators(get()) }
     factory { SetExcludedScanlators(get()) }
+
+    single<RelatedMangaCacheRepository> { RelatedMangaCacheRepositoryImpl(get()) }
+    factory { GetRelatedMangaCache(get()) }
+    factory { SetRelatedMangaCache(get()) }
+    factory { InvalidateRelatedMangaCache(get()) }
 
     factory { SetMangaCategories(get()) }