+++ app/src/main/java/yokai/domain/manga/interactor/GetRelatedMangaCache.kt
@@ -1,0 +1,9 @@
+package yokai.domain.manga.interactor
+
+import yokai.domain.manga.RelatedMangaCacheRepository
+
+class GetRelatedMangaCache(
+    private val relatedMangaCacheRepository: RelatedMangaCacheRepository,
+) {
+    suspend fun await(mangaId: Long): List<Long>? = relatedMangaCacheRepository.getCachedRelatedMangaIds(mangaId)
+}