+++ app/src/main/java/yokai/domain/manga/interactor/InvalidateRelatedMangaCache.kt
@@ -1,0 +1,9 @@
+package yokai.domain.manga.interactor
+
+import yokai.domain.manga.RelatedMangaCacheRepository
+
+class InvalidateRelatedMangaCache(
+    private val relatedMangaCacheRepository: RelatedMangaCacheRepository,
+) {
+    suspend fun await(mangaId: Long) = relatedMangaCacheRepository.invalidate(mangaId)
+}